feat: 验证注册邮箱

+++ server-main/server-main-service/src/main/java/plus/xyc/server/main/account/entity/dto/Account.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author generator
@@ -24,7 +24,7 @@ public class Account implements Serializable {
      * 主键
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 邮箱
